Add inserted sequence for insertions to INFO column
The inserted sequence corresponds to the longest insert in that
location, even if it comes from a CONTROL read.

This is a temporary workaround. I final solution will be provided
by the author of SAVANA, that will output the insertion sequence
to a fasta file, to be referenced in the vcf. This to avoid cluttering
the vcf file with long insertion sequences.